Enable the use of an external helper in order to retrieve a fax URL.
Some fax URLs are dynamic, and cannot be described in
WeatherFaxInternetRetrieval.xml. Finding the charts require (custom) parsing
of the weather provider HTML pages and extracting the correct URL.

This commit delegates the parsing to an external script, which can be
implemented in shell, perl, python, etc.

In order to use an external script, the <Map> XML tag must have a
non NULL "Script" attribute, which is the name of the executable which will
be called by weatherfax in order to find the URL of the fax.

This executable must be placed in the $prefix/share/opencpn/plugins/weatherfax_pi/data/
directory (or the corresponding folder on Windows/Mac). The executable will
take three parameters on the command line:
  - the <Server> "Name" attribute
  - the <Region> "Name" attribute
  - the <Map> "Contents" attribute
The executable must return, on its standard output, on a single line, the
URL to be used by weatherfax.

Example WeatherFaxInternetRetrieval.xml section:
  <Server Name="myserver">
    <Region Name="myregion">
      <Map Contents="mycontents" Area="1" Script="foobar.exe" />
      <Area Name="1" lat1="40N" lat2="70N" lon1="50W" lon2="50E" />
    </Region>
  </Server>

Example foobar.exe:
  #! /bin/sh

  echo http://www.weathercharts.net/noaa_ukmo_analysis/previous/PPVA89_fullsize_00z_last.gif

  exit 0