Fix InternetRetrieval Menu
This fixes the InternetRetrieval Menu so that the the primary selectors Retrieve "Selected" "Selection" are visible when  the menu is opened. This is a huge problem to new users of the plugin and also makes things more difficult for all users trying to adjust the menu so the buttons are visible. (Took 3 stretches to get this done)