src: Use the native apple gl.h header.
The gl.h header is available in all MacOs versions. Since the opengl
headers are designed to be adapted by the OS vendor, they actually do
differ on different platforms. For this reason, using the version
provided by the OS, here MacOS, should be preferred if available.

Adjust the include path on MacOS to the location where gl.h et. al.
lives.